automatically update mining pools list
